+++ docs/app/app.module.ts
@@ -2,8 +2,6 @@ declare var angular: angular.IAngularStatic;
 
 let app = angular.module('app', ['ux-aspects']);
 
-import { IDocumentationPage } from './interfaces/IDocumentationPage';
-
 import { NgModule, forwardRef } from '@angular/core';
 import { HttpModule } from '@angular/http';
 import { FormsModule } from '@angular/forms';
@@ -46,7 +44,6 @@ import { ChartsPageComponent } from './pages/charts/charts.component';
 // import { CustomizePageComponent } from './pages/customize/customize.component';
 import { TeamPageComponent } from './pages/team/team.component';
 import { BlogPageComponent } from './pages/blog/blog.component';
-// import { DocumentationCategoryComponent } from './components/documentation-category/documentation-category.component';
 import { LicensesPageComponent } from './pages/licenses/licenses.component';
 import { ChangeLogPageComponent } from './pages/changelog/changelog.component';
 
@@ -54,26 +51,6 @@ import { documentationSections } from './decorators/documentation-section-compon
 
 
 export const upgradeAdapter = new UpgradeAdapter(forwardRef(() => AppModule));
-
-// Import Page Sections
-// import { CssPageModule } from './pages/css/css.module';
-// import { ComponentsPageModule } from './pages/components/components.module';
-// import { ChartsPageModule } from './pages/charts/charts.module';
-
-/*
-  Configure Component Page Routes
-*/
-// const componentRoutes = loadRoutes(require('./data/components-page.json'));
-
-/*
-  Configure CSS Page Routes
-*/
-// const cssRoutes = loadRoutes(require('./data/css-page.json'));
-
-/*
-  Configure Charts Page Routes
-*/
-// const chartRoutes = loadRoutes(require('./data/charts-page.json'));
 
 /*
   Configure Application Routes
@@ -111,7 +88,6 @@ const DECLARATIONS = [
   // CustomizePageComponent,
   TeamPageComponent,
   BlogPageComponent,
-  // DocumentationCategoryComponent,
   LicensesPageComponent,
   ChangeLogPageComponent,
 
@@ -154,9 +130,6 @@ const DECLARATIONS = [
     DocumentationComponentsModule,
     DocumentationDirectivesModule,
     DocumentationProvidersModule,
-    // CssPageModule,
-    // ChartsPageModule,
-    // ComponentsPageModule,
 
     // Bootstrap Modules
     TabsModule.forRoot(),
@@ -205,17 +178,3 @@ app.config(['$anchorScrollProvider', function ($anchorScrollProvider: angular.IA
 
 // bootstrap the Angular 1 application here
 upgradeAdapter.bootstrap(document.documentElement, ['app']);
-
-
-// // Function to load routes from JSON data
-// function loadRoutes(data: IDocumentationPage): Routes {
-//   let routes: Routes = [];
-//   for (let i = 0; i < data.categories.length; i += 1) {
-//     let category = data.categories[i];
-//     if (routes.length === 0) {
-//       routes.push({ path: '', redirectTo: category.link, pathMatch: 'full' });
-//     }
-//     routes.push({ path: category.link, component: DocumentationCategoryComponent, data: { category: category } });
-//   }
-//   return routes;
-// }
