+++ docs/app/pages/components/sections/buttons/dropdown-ng1/dropdown-ng1.component.ts
@@ -2,7 +2,6 @@ import { Component, ViewEncapsulation } from '@angular/core';
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/dropdown-wrapper.directive';
 
 @Component({
     selector: 'uxd-dropdown-ng1',
