+++ docs/app/pages/components/sections/buttons/floating-action-button-ng1/floating-action-button-ng1.component.ts
@@ -2,7 +2,6 @@ import { Component } from '@angular/core';
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/floating-action-button-wrapper.directive';
 
 @Component({
     selector: 'uxd-floating-action-button-ng1',
