+++ docs/app/pages/components/sections/buttons/pagination-ng1/pagination-ng1.component.ts
@@ -2,7 +2,6 @@ import { Component } from '@angular/core';
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/pagination-wrapper.directive';
 
 @Component({
     selector: 'uxd-pagination-ng1',
