+++ docs/app/pages/components/sections/buttons/single-toggle-button-ng1/single-toggle-button-ng1.component.ts
@@ -2,7 +2,6 @@ import { Component } from '@angular/core';
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/single-toggle-button-wrapper.directive';
 
 @Component({
     selector: 'uxd-single-toggle-button-ng1',
