+++ docs/app/pages/components/sections/buttons/thumbnail-ng1/thumbnail-ng1.component.ts
@@ -2,7 +2,6 @@ import { Component, ViewEncapsulation } from '@angular/core';
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/thumbnail-wrapper.directive';
 
 @Component({
     selector: 'uxd-thumbnail-ng1',
