+++ docs/app/pages/components/sections/component-list/component-list-ng1/component-list-ng1.component.ts
@@ -2,7 +2,6 @@ import { Component, ViewEncapsulation } from '@angular/core';
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/component-list-wrapper.directive';
 
 @Component({
     selector: 'uxd-component-list-ng1',
