+++ docs/app/pages/components/sections/contacts/contacts-ng1/contacts-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/contacts-wrapper.directive';
 
 @Component({
     selector: 'uxd-contacts-ng1',
