+++ docs/app/pages/components/sections/contacts/contacts-overflow-ng1/contacts-overflow-ng1.component.ts
@@ -2,7 +2,6 @@ import { Component, ViewEncapsulation } from '@angular/core';
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/contacts-overflow-wrapper.directive';
 
 @Component({
     selector: 'uxd-contacts-overflow-ng1',
