+++ docs/app/pages/components/sections/draggable-cards/draggable-cards-ng1/draggable-cards-ng1.component.ts
@@ -2,7 +2,6 @@ import { Component, ViewEncapsulation } from '@angular/core';
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/draggable-cards-wrapper.directive';
 
 @Component({
     selector: 'uxd-draggable-cards-ng1',
