+++ docs/app/pages/components/sections/draggable-panels/draggable-panels-views-ng1/draggable-panels-views-ng1.component.ts
@@ -2,7 +2,6 @@ import { Component, ViewEncapsulation } from '@angular/core';
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/draggable-panels-views-wrapper.directive';
 
 @Component({
     selector: 'uxd-draggable-panels-views-ng1',
