+++ docs/app/pages/components/sections/keyboard/hotkeys-ng1/hotkeys-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/hotkeys-wrapper.directive';
 
 @Component({
     selector: 'uxd-hotkeys-ng1',
