+++ docs/app/pages/components/sections/keyboard/keyboard-service-ng1/keyboard-service-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/keyboard-service-wrapper.directive';
 
 @Component({
     selector: 'uxd-keyboard-service-ng1',
