+++ docs/app/pages/components/sections/modals/modal-ng1/modal-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/modal-wrapper.directive.js';
 
 @Component({
     selector: 'uxd-modal-ng1',
