+++ docs/app/pages/components/sections/modals/side-modal-ng1/side-modal-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/side-modal-wrapper.directive.js';
 
 @Component({
     selector: 'uxd-side-modal-ng1',
