+++ docs/app/pages/components/sections/modals/square-modal-ng1/square-modal-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/square-modal-wrapper.directive.js';
 
 @Component({
     selector: 'uxd-square-modal-ng1',
