+++ docs/app/pages/components/sections/notifications/alert-styles-ng1/alert-styles-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/alert-styles-wrapper.directive.js';
 
 @Component({
     selector: 'uxd-alert-styles-ng1',
