+++ docs/app/pages/components/sections/notifications/dismissable-styles-ng1/dismissable-styles-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/dismissable-styles-wrapper.directive.js';
 
 @Component({
     selector: 'uxd-dismissable-styles-ng1',
