+++ docs/app/pages/components/sections/notifications/notifications-ng1/notifications-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/notifications-wrapper.directive.js';
 
 @Component({
     selector: 'uxd-notifications-ng1',
