+++ docs/app/pages/components/sections/panels/collapsible-panels-ng1/collapsible-panels-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/collapsible-panels-wrapper.directive.js';
 
 @Component({
     selector: 'uxd-collapsible-panels-ng1',
