+++ docs/app/pages/components/sections/panels/item-display-panel-ng1/item-display-panel-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/item-display-panel-wrapper.directive.js';
 
 @Component({
     selector: 'uxd-item-display-panel-ng1',
