+++ docs/app/pages/components/sections/panels/modal-inset-panel-ng1/item-display-panel-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/modal-inset-panel-wrapper.directive.js';
 
 @Component({
     selector: 'uxd-modal-inset-panel-ng1',
