+++ docs/app/pages/components/sections/popover/popover-ng1/popover-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/popover-wrapper.directive.js';
 
 @Component({
     selector: 'uxd-popover-ng1',
