+++ docs/app/pages/components/sections/progress/progress-bar-ng1/progress-bar-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/progress-bar-wrapper.directive.js';
 
 @Component({
     selector: 'uxd-progress-bar-ng1',
