+++ docs/app/pages/components/sections/scrollbar/custom-scrollbar-ng1/custom-scrollbar-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/custom-scrollbar-wrapper.directive.js';
 
 @Component({
     selector: 'uxd-custom-scrollbar-ng1',
