+++ docs/app/pages/components/sections/scrollbar/infinite-scroll-ng1/infinite-scroll-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/infinite-scroll-wrapper.directive.js';
 
 @Component({
     selector: 'uxd-infinite-scroll-ng1',
