+++ docs/app/pages/components/sections/search/search-history-ng1/search-history-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/search-history-wrapper.directive.js';
 
 @Component({
     selector: 'uxd-search-history-ng1',
