+++ docs/app/pages/components/sections/search/search-toolbar-ng1/search-toolbar-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/search-toolbar-wrapper.directive.js';
 
 @Component({
     selector: 'uxd-search-toolbar-ng1',
