+++ docs/app/pages/components/sections/splitter/nested-splitter-ng1/nested-splitter-ng1.component.ts
@@ -3,7 +3,6 @@ import { BaseDocumentationSection } from '../../../../../components/base-documen
 import { ICodePenProvider } from '../../../../../interfaces/ICodePenProvider';
 import { ICodePen } from '../../../../../interfaces/ICodePen';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
-import './wrapper/nested-splitter-wrapper.directive.js';
 
 @Component({
     selector: 'uxd-nested-splitter-ng1',
