+++ docs/app/pages/css/css.module.ts
@@ -93,15 +93,7 @@ import { CssDynamicNameCalloutComponent } from './sections/page-header/dynamic-n
 import { CssStandardHeaderToolbarComponent } from './sections/page-header/standard-header-toolbar/standard-header-toolbar.component';
 import { CssHeaderNavTabToolbarComponent } from './sections/page-header/header-nav-tab-toolbar/header-nav-tab-toolbar.component';
 import { DocumentationCategoryComponent } from '../../components/documentation-category/documentation-category.component';
-
-// Import Wrappers
-import './sections/forms/form-validation-field-by-field/wrapper/form-validation-field-by-field-wrapper.directive';
-import './sections/forms/form-validation-on-submit/wrapper/form-validation-on-submit-wrapper.directive';
-import './sections/text-inputs/float-labels/wrapper/float-labels-wrapper.directive';
-import './sections/side-navigation/navigation/wrapper/navigation-wrapper.directive';
-import './sections/side-navigation/navigation/wrapper/boldify.directive';
-import './sections/side-navigation/app-navigator/wrapper/app-navigator-wrapper.directive';
-
+import { WrappersModule } from '../../wrappers.module';
 
 const CSS_SECTIONS = [
     CssColoredButtonsComponent,
@@ -192,8 +184,6 @@ const CSS_SECTIONS = [
     upgradeAdapter.upgradeNg1Component('uxdAppNavigatorWrapper'),
 ];
 
-const cssRoutes = loadRoutes(require('../../data/css-page.json'));
-
 @NgModule({
     imports: [
         TabsModule,
@@ -202,24 +192,17 @@ const cssRoutes = loadRoutes(require('../../data/css-page.json'));
         DocumentationDirectivesModule,
         FormsModule,
         CommonModule,
-        RouterModule.forChild(cssRoutes)
+        WrappersModule,
+        RouterModule.forChild(ResolverService.resolveRouteComponents(require('../../data/css-page.json')))
     ],
     exports: CSS_SECTIONS,
     declarations: CSS_SECTIONS,
     entryComponents: CSS_SECTIONS,
     providers: [],
 })
-export class CssPageModule { }
+export class CssPageModule {
 
-// Function to load routes from JSON data
-function loadRoutes(data: IDocumentationPage): Routes {
-    let routes: Routes = [];
-    for (let i = 0; i < data.categories.length; i += 1) {
-        let category = data.categories[i];
-        if (routes.length === 0) {
-            routes.push({ path: '', redirectTo: category.link, pathMatch: 'full' });
-        }
-        routes.push({ path: category.link, component: DocumentationCategoryComponent, data: { category: category } });
+    constructor(componentFactoryResolver: ComponentFactoryResolver, resolverService: ResolverService) {
+        resolverService.registerResolver(componentFactoryResolver);
     }
-    return routes;
 }
