+++ docs/app/services/edit-example/edit-example.service.ts
@@ -1,5 +1,4 @@
-import { Injectable, Inject } from '@angular/core';
-import { DOCUMENT } from '@angular/platform-browser';
+import { Injectable } from '@angular/core';
 
 import { CodePenService } from '../codepen/codepen.service';
 import { PlunkerService } from '../plunker/plunker.service';
@@ -11,7 +10,7 @@ export class EditExampleService {
 
     constructor(private codePenService: CodePenService, private plunkerService: PlunkerService) {}
 
-    public launchEditor(title: string, content: any, type: 'codepen' | 'plunker', version: 'Angular' | 'AngularJS') {
+    public launchEditor(title: string, content: any, type: 'codepen' | 'plunker') {
         switch (type) {
             case 'codepen':
                 this.codePenService.launch(title, <ICodePen>content);
