+++ tsconfig.json
@@ -15,9 +15,6 @@
     "debug": true,
     "skipMetadataEmit": true
   },
-  "awesomeTypescriptLoaderOptions": {
-    "useWebpackText": true
-  },
   "exclude": [
     "node_modules",
     "dist"
